# contracts/openapi.yaml
components:
  schemas:
    AppliedPromotionOut:
      properties:
        code:
          title: Code
          type: string
        discount_vnd:
          title: Discount Vnd
          type: integer
        name:
          title: Name
          type: string
        terms_url:
          anyOf:
          - type: string
          - type: 'null'
          title: Terms Url
      required:
      - code
      - name
      - discount_vnd
      - terms_url
      title: AppliedPromotionOut
      type: object
    ColorOut:
      properties:
        code:
          title: Code
          type: string
        exterior:
          title: Exterior
          type: boolean
        name:
          title: Name
          type: string
      required:
      - code
      - name
      - exterior
      title: ColorOut
      type: object
    CompensationOut:
      description: Số tiền đòi khách trả, nên trả kèm toàn bộ căn cứ đã dùng.
      properties:
        compensation_vnd:
          title: Compensation Vnd
          type: integer
        floor_applied:
          title: Floor Applied
          type: boolean
        list_price_as_of:
          format: date
          title: List Price As Of
          type: string
        list_price_vnd:
          title: List Price Vnd
          type: integer
        months_used:
          title: Months Used
          type: integer
        warranty_months:
          title: Warranty Months
          type: integer
        warranty_policy_as_of:
          format: date
          title: Warranty Policy As Of
          type: string
      required:
      - compensation_vnd
      - list_price_vnd
      - list_price_as_of
      - months_used
      - warranty_months
      - warranty_policy_as_of
      - floor_applied
      title: CompensationOut
      type: object
    CreateOrderIn:
      properties:
        as_of:
          anyOf:
          - format: date
            type: string
          - type: 'null'
          title: As Of
        code:
          title: Code
          type: string
        deposit_vnd:
          title: Deposit Vnd
          type: integer
        ownership_model:
          $ref: '#/components/schemas/OwnershipModel'
          default: battery_included
        province_code:
          title: Province Code
          type: string
        trim_id:
          title: Trim Id
          type: integer
      required:
      - code
      - trim_id
      - province_code
      - deposit_vnd
      title: CreateOrderIn
      type: object
    FuelType:
      enum:
      - gasoline
      - diesel
      title: FuelType
      type: string
    HTTPValidationError:
      properties:
        detail:
          items:
            $ref: '#/components/schemas/ValidationError'
          title: Detail
          type: array
      title: HTTPValidationError
      type: object
    LineItemOut:
      properties:
        amount_vnd:
          title: Amount Vnd
          type: integer
        code:
          title: Code
          type: string
        label:
          title: Label
          type: string
      required:
      - code
      - label
      - amount_vnd
      title: LineItemOut
      type: object
    ModelOut:
      properties:
        category:
          title: Category
          type: string
        code:
          title: Code
          type: string
        colors:
          items:
            $ref: '#/components/schemas/ColorOut'
          title: Colors
          type: array
        id:
          title: Id
          type: integer
        name:
          title: Name
          type: string
        seats:
          anyOf:
          - type: integer
          - type: 'null'
          title: Seats
        segment:
          anyOf:
          - type: string
          - type: 'null'
          title: Segment
        trims:
          items:
            $ref: '#/components/schemas/TrimOut'
          title: Trims
          type: array
      required:
      - id
      - code
      - name
      - category
      - segment
      - seats
      - trims
      - colors
      title: ModelOut
      type: object
    OnRoadQuoteOut:
      properties:
        as_of:
          format: date
          title: As Of
          type: string
        discount_vnd:
          title: Discount Vnd
          type: integer
        items:
          items:
            $ref: '#/components/schemas/LineItemOut'
          title: Items
          type: array
        monthly_battery_fee_vnd:
          anyOf:
          - type: integer
          - type: 'null'
          title: Monthly Battery Fee Vnd
        ownership_model:
          title: Ownership Model
          type: string
        promotions:
          items:
            $ref: '#/components/schemas/AppliedPromotionOut'
          title: Promotions
          type: array
        province_code:
          title: Province Code
          type: string
        total_vnd:
          title: Total Vnd
          type: integer
        vehicle_price_vnd:
          title: Vehicle Price Vnd
          type: integer
      required:
      - as_of
      - province_code
      - ownership_model
      - vehicle_price_vnd
      - discount_vnd
      - monthly_battery_fee_vnd
      - items
      - promotions
      - total_vnd
      title: OnRoadQuoteOut
      type: object
    OrderOut:
      properties:
        channel:
          title: Channel
          type: string
        code:
          title: Code
          type: string
        deposit_vnd:
          title: Deposit Vnd
          type: integer
        price_snapshot:
          additionalProperties: true
          title: Price Snapshot
          type: object
        province_code:
          title: Province Code
          type: string
        quoted_on:
          format: date
          title: Quoted On
          type: string
        status:
          title: Status
          type: string
        total_vnd:
          title: Total Vnd
          type: integer
        trim_id:
          title: Trim Id
          type: integer
      required:
      - code
      - status
      - channel
      - trim_id
      - province_code
      - quoted_on
      - total_vnd
      - deposit_vnd
      - price_snapshot
      title: OrderOut
      type: object
    OwnershipModel:
      description: Hai mô hình sở hữu pin. Đây là hai biến thể giá, không phải một cờ.
      enum:
      - battery_included
      - battery_lease
      title: OwnershipModel
      type: string
    PromotionOut:
      properties:
        code:
          title: Code
          type: string
        effective_from:
          format: date
          title: Effective From
          type: string
        effective_to:
          anyOf:
          - format: date
            type: string
          - type: 'null'
          title: Effective To
        kind:
          title: Kind
          type: string
        name:
          title: Name
          type: string
        terms_url:
          anyOf:
          - type: string
          - type: 'null'
          title: Terms Url
      required:
      - code
      - name
      - kind
      - terms_url
      - effective_from
      - effective_to
      title: PromotionOut
      type: object
    RunningCostOut:
      properties:
        as_of:
          format: date
          title: As Of
          type: string
        electricity_price_as_of:
          format: date
          title: Electricity Price As Of
          type: string
        ev_monthly_cost_vnd:
          title: Ev Monthly Cost Vnd
          type: integer
        fuel_price_as_of:
          format: date
          title: Fuel Price As Of
          type: string
        ice_monthly_cost_vnd:
          title: Ice Monthly Cost Vnd
          type: integer
        monthly_km:
          title: Monthly Km
          type: integer
        monthly_saving_vnd:
          title: Monthly Saving Vnd
          type: integer
      required:
      - as_of
      - monthly_km
      - ev_monthly_cost_vnd
      - ice_monthly_cost_vnd
      - monthly_saving_vnd
      - electricity_price_as_of
      - fuel_price_as_of
      title: RunningCostOut
      type: object
    SpecOut:
      description: Thông số có kiểu. `ai/` đọc đúng schema này, không đọc văn bản mô tả.
      properties:
        drivetrain:
          anyOf:
          - type: string
          - type: 'null'
          title: Drivetrain
        fast_charge_minutes:
          anyOf:
          - type: integer
          - type: 'null'
          title: Fast Charge Minutes
        max_power_w:
          anyOf:
          - type: integer
          - type: 'null'
          title: Max Power W
        max_torque_nm:
          anyOf:
          - type: integer
          - type: 'null'
          title: Max Torque Nm
        motor_count:
          anyOf:
          - type: integer
          - type: 'null'
          title: Motor Count
        range_nedc_km:
          anyOf:
          - type: integer
          - type: 'null'
          title: Range Nedc Km
        range_per_charge_km:
          anyOf:
          - type: integer
          - type: 'null'
          title: Range Per Charge Km
        top_speed_kmh:
          anyOf:
          - type: integer
          - type: 'null'
          title: Top Speed Kmh
      title: SpecOut
      type: object
    TrimOut:
      properties:
        code:
          title: Code
          type: string
        id:
          title: Id
          type: integer
        name:
          title: Name
          type: string
        spec:
          anyOf:
          - $ref: '#/components/schemas/SpecOut'
          - type: 'null'
      required:
      - id
      - code
      - name
      title: TrimOut
      type: object
    ValidationError:
      properties:
        ctx:
          title: Context
          type: object
        input:
          title: Input
        loc:
          items:
            anyOf:
            - type: string
            - type: integer
          title: Location
          type: array
        msg:
          title: Message
          type: string
        type:
          title: Error Type
          type: string
      required:
      - loc
      - msg
      - type
      title: ValidationError
      type: object
info:
  description: Nền tảng web bán ô tô điện.
  title: IFast API
  version: 0.1.0
openapi: 3.1.0
paths:
  /battery/assets/{serial}/compensation:
    get:
      operationId: compensation_for_asset_battery_assets__serial__compensation_get
      parameters:
      - in: path
        name: serial
        required: true
        schema:
          title: Serial
          type: string
      - in: query
        name: incident_on
        required: true
        schema:
          format: date
          title: Incident On
          type: string
      responses:
        '200':
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/CompensationOut'
          description: Successful Response
        '422':
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/HTTPValidationError'
          description: Validation Error
      summary: Compensation For Asset
      tags:
      - battery
  /catalog/models:
    get:
      operationId: list_models_catalog_models_get
      responses:
        '200':
          content:
            application/json:
              schema:
                items:
                  $ref: '#/components/schemas/ModelOut'
                title: Response List Models Catalog Models Get
                type: array
          description: Successful Response
      summary: List Models
      tags:
      - catalog
  /catalog/models/{code}:
    get:
      operationId: get_model_catalog_models__code__get
      parameters:
      - in: path
        name: code
        required: true
        schema:
          title: Code
          type: string
      responses:
        '200':
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ModelOut'
          description: Successful Response
        '422':
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/HTTPValidationError'
          description: Validation Error
      summary: Get Model
      tags:
      - catalog
  /health:
    get:
      operationId: health_health_get
      responses:
        '200':
          content:
            application/json:
              schema:
                additionalProperties:
                  type: string
                title: Response Health Health Get
                type: object
          description: Successful Response
      summary: Health
      tags:
      - system
  /orders:
    post:
      description: 'Tạo đơn và khoá snapshot.


        Sau lời gọi này, đơn không còn phụ thuộc vào bảng giá. Bảng giá đổi ngày

        mai không làm đổi con số trong đơn.'
      operationId: create_orders_post
      requestBody:
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/CreateOrderIn'
        required: true
      responses:
        '201':
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/OrderOut'
          description: Successful Response
        '422':
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/HTTPValidationError'
          description: Validation Error
      summary: Create
      tags:
      - orders
  /orders/{code}:
    get:
      operationId: get_order_orders__code__get
      parameters:
      - in: path
        name: code
        required: true
        schema:
          title: Code
          type: string
      responses:
        '200':
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/OrderOut'
          description: Successful Response
        '422':
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/HTTPValidationError'
          description: Validation Error
      summary: Get Order
      tags:
      - orders
  /pricing/on-road:
    get:
      operationId: on_road_quote_pricing_on_road_get
      parameters:
      - in: query
        name: trim_id
        required: true
        schema:
          title: Trim Id
          type: integer
      - in: query
        name: province_code
        required: true
        schema:
          title: Province Code
          type: string
      - in: query
        name: ownership_model
        required: false
        schema:
          $ref: '#/components/schemas/OwnershipModel'
          default: battery_included
      - description: Mốc thời gian tính giá. Bỏ trống thì lấy hôm nay; báo giá cho một đơn cũ phải truyền
          đúng ngày của đơn đó.
        in: query
        name: as_of
        required: false
        schema:
          anyOf:
          - format: date
            type: string
          - type: 'null'
          description: Mốc thời gian tính giá. Bỏ trống thì lấy hôm nay; báo giá cho một đơn cũ phải truyền
            đúng ngày của đơn đó.
          title: As Of
      responses:
        '200':
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/OnRoadQuoteOut'
          description: Successful Response
        '422':
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/HTTPValidationError'
          description: Validation Error
      summary: On Road Quote
      tags:
      - pricing
  /pricing/running-cost:
    get:
      operationId: running_cost_pricing_running_cost_get
      parameters:
      - in: query
        name: monthly_km
        required: true
        schema:
          title: Monthly Km
          type: integer
      - in: query
        name: ev_consumption_kwh_per_100km_x10
        required: true
        schema:
          title: Ev Consumption Kwh Per 100Km X10
          type: integer
      - in: query
        name: ice_consumption_litres_per_100km_x10
        required: true
        schema:
          title: Ice Consumption Litres Per 100Km X10
          type: integer
      - in: query
        name: fuel_type
        required: false
        schema:
          $ref: '#/components/schemas/FuelType'
          default: gasoline
      - in: query
        name: as_of
        required: false
        schema:
          anyOf:
          - format: date
            type: string
          - type: 'null'
          title: As Of
      responses:
        '200':
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/RunningCostOut'
          description: Successful Response
        '422':
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/HTTPValidationError'
          description: Validation Error
      summary: Running Cost
      tags:
      - pricing
  /promotions:
    get:
      description: 'Ưu đãi còn hiệu lực tại một mốc thời gian.


        Mặc định là hôm nay, nhưng tra cứu lại một đơn cũ phải truyền đúng ngày của

        đơn — ưu đãi đã hết hạn vẫn phải giải thích được vì sao từng được áp.'
      operationId: list_promotions_promotions_get
      parameters:
      - in: query
        name: as_of
        required: false
        schema:
          anyOf:
          - format: date
            type: string
          - type: 'null'
          title: As Of
      responses:
        '200':
          content:
            application/json:
              schema:
                items:
                  $ref: '#/components/schemas/PromotionOut'
                title: Response List Promotions Promotions Get
                type: array
          description: Successful Response
        '422':
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/HTTPValidationError'
          description: Validation Error
      summary: List Promotions
      tags:
      - promotions